refactor(data): share sqlite materialization
