+++ src/data/types.ts
@@ -90,8 +90,6 @@ export interface Entry {
   goodlift: number | null;
 }
 
-// One row per (category, sex, equipmentGroup, weightClass, rank). Precomputed
-// in the SQLite snapshot; ~17k rows total across all combinations.
 export type RecordCategory =
   | "squat_full_power"
   | "squat_all_events"
@@ -103,24 +101,6 @@ export type RecordCategory =
 
 export type EquipmentGroup = "raw" | "wraps" | "single" | "multi" | "unlimited" | "all-tested";
 
-export interface WeightClassRecord {
-  category: RecordCategory;
-  sex: Sex;
-  equipmentGroup: EquipmentGroup;
-  weightClassKg: number;
-  rank: number; // 1..3
-  entryId: number; // references entries.id or fixture entries[]
-  liftValue: number;
-}
-
-// /api/federations row shape. Materialized into the SQLite snapshot.
-export interface FederationSummary {
-  slug: string; // lowercased + alphanumeric-only ("wrpfuk")
-  code: string; // original casing ("WRPF-UK")
-  parentSlug: string | null;
-  meetCount: number;
-}
-
 // Ranking metric keys. Each maps to an Entry field and a SQLite ranking column.
 export type RankMetric =
   | "dots"
@@ -131,36 +111,3 @@ export type RankMetric =
   | "squat"
   | "bench"
   | "deadlift";
-
-export type MetricInt32Arrays = Record<RankMetric, Int32Array>;
-export type MetricUint32Arrays = Record<RankMetric, Uint32Array>;
-
-export interface AppData {
-  // Entity arrays — index in the array IS the canonical id.
-  lifters: Lifter[];
-  meets: Meet[];
-  entries: Entry[];
-
-  // Lookup maps.
-  lifterByUsername: Map<string, number>;
-  meetByPath: Map<string, number>;
-  entriesByLifter: Map<number, number[]>; // lifterId → [entryId, ...]
-  entriesByMeet: Map<number, number[]>; // meetId → [entryId, ...]
-
-  // Per-lifter best entry per ranking metric. -1 = no eligible entry.
-  bestEntryByLifter: MetricInt32Arrays;
-  // Lifter ids sorted DESC by their best on that metric. Top-N = .subarray(0, N).
-  rankByMetric: MetricUint32Arrays;
-
-  // Records leaderboard (top-3 per category/sex/equipment_group/weight_class).
-  records: WeightClassRecord[];
-
-  // Federations index for /api/federations.
-  federations: FederationSummary[];
-  meetsByFederation: Map<string, number[]>; // slug → [meetId, ...]
-
-  // Metadata about this snapshot.
-  sourceLastModified: string | null;
-  ingestedAt: string; // ISO 8601
-  rowCount: number;
-}
